requirements.txt
filedate==2.0
google_api_python_client==2.83.0
httplib2==0.20.4
oauth2client==4.1.3
protobuf==4.22.1
requests==2.28.0
